refactor(layering): subpackage publishing release and metadata families
Phase 2b of RENDERING-LAYERING-1 (row closed). The five release_*
modules moved to infrastructure/publishing/release/, the four
metadata_* modules to infrastructure/publishing/metadata/, and the
flat metadata.py aggregator became metadata/metadata_aggregate.py
with the historical module-path imports resolving through the
package __init__.

- Silent re-export shims keep all nine flat paths plus the
  python -m infrastructure.publishing.metadata_export_cli execution
  path working.
- Every consumer retargeted to canonical nested paths: publishing
  internals (workflow, zenodo helpers, ledger, rehearsal,
  publishing/__init__), the drift gate, Stage-12 pipeline,
  transmission bookends, thin scripts, five test files, and the
  Zenodo DOI strategy guide link.
- New subpackages carry AGENTS.md/README.md doc pairs; publishing
  AGENTS/README/SKILL tables and examples updated; metadata SKILL.md
  usage example corrected to the real generate_metadata_package
  signature.
- Layering guard still runs with an empty allowlist; api-reference
  stays at 26 top-level packages; counts refreshed with provenance.
