+++ infrastructure/publishing/release/__init__.py
@@ -1,0 +1,6 @@
+"""Release receipt/pairing/workflow family for the publishing pipeline.
+
+Subpackaged out of the flat ``infrastructure.publishing`` namespace by
+RENDERING-LAYERING-1 phase 2b; the flat module paths remain as silent
+backwards-compat shims.
+"""
