+++ scripts/pipeline/stage_12_metadata.py
@@ -12,7 +12,7 @@ from scripts import ensure_repo_root_on_path  # noqa: E402
 
 ensure_repo_root_on_path()
 
-from infrastructure.publishing.metadata_stage import run_metadata_package  # noqa: E402
+from infrastructure.publishing.metadata.metadata_stage import run_metadata_package  # noqa: E402
 
 
 def _parse_args() -> argparse.Namespace:
